Prepare projects for conjoining
