+++ SharpScapeClient/client/scenes/LoginModal.cs
