+++ SharpScapeClient/client/scenes/LoginModal.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=3 format=2]
 
-[ext_resource path="res://client/LoginModal.cs" type="Script" id=1]
+[ext_resource path="res://client/scenes/LoginModal.cs" type="Script" id=1]
 
 [sub_resource type="StyleBoxFlat" id=1]
 bg_color = Color( 0.25098, 0.145098, 0.145098, 1 )
