+++ SharpScapeClient/client/scripts/ApiPayloadSecurity.cs
