+++ SharpScapeClient/project.godot
@@ -11,7 +11,7 @@ config_version=4
 [application]
 
 config/name="SharpScapeClient"
-run/main_scene="res://client/client.tscn"
+run/main_scene="res://client/scenes/ClientUI.tscn"
 config/icon="res://icon.png"
 
 [autoload]
