+++ SharpScapeServer/project.godot
@@ -11,7 +11,7 @@ config_version=4
 [application]
 
 config/name="SharpScapeServer"
-run/main_scene="res://server/server.tscn"
+run/main_scene="res://server/scenes/server.tscn"
 config/icon="res://icon.png"
 
 [autoload]
