+++ SharpScapeServer/server/scenes/server.tscn
@@ -1,15 +1,15 @@
 [gd_scene load_steps=3 format=2]
 
-[ext_resource path="res://server/Server.cs" type="Script" id=1]
-[ext_resource path="res://server/ServerUI.cs" type="Script" id=2]
+[ext_resource path="res://server/scenes/ServerUI.cs" type="Script" id=1]
+[ext_resource path="res://server/scripts/SharpScapeServer.cs" type="Script" id=2]
 
 [node name="ServerControl" type="Control"]
 anchor_right = 1.0
 anchor_bottom = 1.0
-script = ExtResource( 2 )
-
-[node name="Server" type="Node" parent="."]
 script = ExtResource( 1 )
+
+[node name="SharpScapeServer" type="Node" parent="."]
+script = ExtResource( 2 )
 
 [node name="Panel" type="Panel" parent="."]
 anchor_right = 1.0
