+++ SharpScapeServer/server/scripts/Http.cs
