+++ SharpScapeServer/server/scripts/MPServerCrypto.cs
